refactor: extract frontend moderation service
