+++ apps/client/src/app/data-access/chat/chat.service.interface.ts
@@ -1,12 +1,4 @@
-import {
-  type BanUserDto,
-  type ChatMessage,
-  type GetMessagesResponseDto,
-  type ModerationQueueDto,
-  type ModerationSummaryDto,
-  type ReportChatMessageDto,
-  type UpdateChatReportStatusDto
-} from "@fuzzy-waddle/api-interfaces";
+import { type ChatMessage, type GetMessagesResponseDto, type ReportChatMessageDto } from "@fuzzy-waddle/api-interfaces";
 import { Observable } from "rxjs";
 
 export interface IChatService {
@@ -14,9 +6,4 @@ export interface IChatService {
   getMessageListener(): Promise<Observable<ChatMessage> | undefined>;
   getMessages(limit?: number, offset?: number, gameInstanceId?: string): Promise<GetMessagesResponseDto>;
   reportMessage(messageId: number, report: ReportChatMessageDto): Promise<void>;
-  getModerationSummary(): Promise<ModerationSummaryDto>;
-  getModerationReports(): Promise<ModerationQueueDto>;
-  updateReportStatus(reportId: number, body: UpdateChatReportStatusDto): Promise<void>;
-  banUser(userId: string, body: BanUserDto): Promise<void>;
-  unbanUser(userId: string): Promise<void>;
 }
